+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # main
 
 - Integrate Node bindings from [`rescript-nodejs`](https://github.com/TheSpyder/rescript-nodejs).
+- Bind `Transpile` and `Build`.
 
 # 0.2.1